# Avoid unnecessary time.Since() when ending spans if the endTime is already specified