Hi, 

When ending a span, I noticed that there might be an unnecessary and expensive `time.Since` invocation (inside the `internal.MonotonicEndTime`) that we can avoid:

Saying that I want to instrument the trace and the latency metric at the same time, I would like to make sure I use only two expensive `time.Now` invocations:
```go
ts1 := time.Now()
ctx, span := tracer.Start(ctx, "myFunc", trace.WithTimestamp(ts1)) 

myFunc(ctx, ...)

ts2 := time.Now()
span.End(trace.WithTimestamp(ts2))
latency.Record(ctx, ts2.Sub(ts1).Milliseconds())
```

So, if there is a `trace.WithTimestamp(ts)` passed in while ending the span, we don't need the `internal.MonotonicEndTime` invocation.

To do that, this PR moves `config := trace.NewSpanEndConfig(options...)` forward and check if there is a `trace.WithTimestamp(ts)` passed in and call `internal.MonotonicEndTime` only when necessary.