@griggsca91 2024-08-01
I'm not sure if this is still an issue that is wanting to be done, but I did a small benchmark and got these results with this change (with a much more recent branch)

```
awesomeProject benchstat old.txt new.txt
goos: darwin
goarch: arm64
pkg: awesomeProject
                                            │   old.txt   │              new.txt               │
                                            │   sec/op    │   sec/op     vs base               │
SpanEnd/span_end_with_timestamp_config-8      726.9n ± 1%   684.0n ± 2%  -5.90% (p=0.000 n=20)
SpanEnd/span_end_without_timestamp_config-8   662.6n ± 4%   672.1n ± 2%       ~ (p=0.147 n=20)
geomean                                       694.0n        678.0n       -2.30%
```

This is the benchmark I used

https://go.dev/play/p/--mM4hpzckk



Is this something we want to try to proceed with for some small savings?